+++ lib/src/array2d/int/int_fix_array2d.dart
@@ -27,7 +27,7 @@ class Int2DFix extends Object
 
   Int2DFix.shaped(Index2D shape, {int data: 0})
       : _data = new List<Int1D>.generate(
-            shape.row, (_) => new Int1D.sized(shape.column, data: data));
+            shape.row, (_) => new Int1D.sized(shape.col, data: data));
 
   factory Int2DFix.shapedLike(Array2DView like, {int data: 0}) =>
       new Int2DFix.sized(like.numRows, like.numCols, data: data);